Improved protection to type of final_insample_loss, to protect from numerical errors

+++ nixtla/models/nbeats/nbeats.py
@@ -440,11 +440,11 @@ class Nbeats(object):
                 lr_scheduler.step()
                 if (iteration % eval_steps == 0):
                     display_string = 'Step: {}, Time: {:03.3f}, Insample {}: {:.5f}'.format(iteration,
-                                                                                    time.time()-start,
-                                                                                    self.loss,
-                                                                                    training_loss.cpu().data.numpy())
+                                                                                            time.time()-start,
+                                                                                            self.loss,
+                                                                                            training_loss.cpu().data.numpy())
                     self.trajectories['iteration'].append(iteration)
-                    self.trajectories['train_loss'].append(training_loss.cpu().data.numpy())
+                    self.trajectories['train_loss'].append(np.float(training_loss.cpu().data.numpy()))
 
                     if val_ts_loader is not None:
                         loss = self.evaluate_performance(ts_loader=val_ts_loader,
@@ -475,7 +475,7 @@ class Nbeats(object):
 
         #End of fitting
         if n_iterations >0:
-            self.final_insample_loss = training_loss.cpu().data.numpy() if not break_flag else best_insample_loss #This is batch!
+            self.final_insample_loss = np.float(training_loss.cpu().data.numpy()) if not break_flag else best_insample_loss #This is batch!
             string = 'Step: {}, Time: {:03.3f}, Insample {}: {:.5f}'.format(iteration,
                                                                             time.time()-start,
                                                                             self.loss,
